fixing some issues

+++ src/config/db.ts
@@ -21,7 +21,7 @@ const initDB = async () => {
         vehicle_name VARCHAR(100) NOT NULL,
         type VARCHAR(20) NOT NULL CHECK(type IN ('car','bike','van','SUV')),
         registration_number VARCHAR(50) NOT NULL UNIQUE,
-        daily_rent_price NUMERIC(10,2) NOT NULL CHECK(daily_rent_price>0),
+        daily_rent_price INT NOT NULL CHECK(daily_rent_price>0),
         availability_status VARCHAR(20) NOT NULL CHECK (availability_status IN ('available','booked'))
         )
         `);
